Update README.md
Added TDD Workflow image. The idea here is that you can create your goal first (just enter plaintext), then diff it against the output of your program and iterate the source until the diff is nil. Ideally, this can be automatically converted to automated testcases by Conception in the future.